Added possibility to deploy from IPFS hash, near-mainnet has filters now

+++ deploy.sh
@@ -10,11 +10,14 @@ graph="node_modules/.bin/graph"
 build=
 clean=
 suffix=
+using_path=true
+subgraphName=
+subgraphHash=
 
 main() {
   pushd "$ROOT" &> /dev/null
 
-  while getopts "hbcs:" opt; do
+  while getopts "hbcs" opt; do
     case $opt in
       h) usage && exit 0;;
       b) build=true;;
@@ -25,25 +28,40 @@ main() {
   done
   shift $((OPTIND-1))
 
-  rm -rf build
-
   path="$1"; shift
-  if [[ ! -f "./$path" ]]; then
-    usage_error "Invalid subgraph name, valid names are: `subgraphs`"
+  if [[ "$path" =~ [^@]+@Qm.+ ]]; then
+    using_path=""
+    using_name="true"
+    subgraphName=`printf "$path" | sed -E "s/@Qm.+$//"`
+    subgraphHash=`printf "$path" | sed -E "s/^[^@]+@//"`
+  else
+    if [[ ! -f "./$path" ]]; then
+      usage_error "Invalid subgraph name, valid names are: `subgraphs`"
+    fi
+
+    subgraphName="`deployment_name $path $suffix`"
   fi
 
-  if [[ $build == true ]]; then
-    ./build.sh "$path"
+  if [[ "$using_path" == "true" ]]; then
+    rm -rf build
+
+    if [[ $build == true ]]; then
+      ./build.sh "$path"
+    fi
   fi
 
   if [[ $clean == true ]]; then
-    graphman_remove "`deployment_name $path $suffix`"
+    graphman_remove "$subgraphName"
   fi
 
-  deploy "$path"
+  if [[ "$using_path" == "true" ]]; then
+    deploy_from_manifest "$path"
+  else
+    deploy_from_hash "$subgraphHash" "$subgraphName"
+  fi
 }
 
-deploy() {
+deploy_from_manifest() {
   path="$1"
 
   chain=`chain $path`
@@ -52,6 +70,17 @@ deploy() {
 
   $graph create "$deployment" --node "$node"
   $graph deploy --node $node --ipfs http://127.0.0.1:5001 --version-label v0.0.1 "$deployment" "$path"
+}
+
+deploy_from_hash() {
+  hash="$1"
+  name="$2"
+  jsonRpcCurl="curl -H 'Content-Type: application/json'"
+
+  node="http://127.0.0.1:8020"
+
+  $graph create "$name" --node "$node"
+  curl -H 'Content-Type: application/json' -d "{\"id\":\"2.0\",\"method\":\"subgraph_deploy\",\"params\":{\"name\":\"$name\", \"ipfs_hash\": \"$hash\", \"version_label\": \"v0.0.1\", \"debug_fork\": \"\" }}" "$node"
 }
 
 usage() {
